9250 - Deleting multiple rules at EOF attempts to select off end of doc
When we remove multiple rules, we attempt to select the next rule in
the range if such a rule exists. Due to the ordering of the selection
update versus the ordering of the `change` listeners we can end up
with ranges in the multi range item list whose actual content has
already been removed!

This means when we want to move to the next element of the list, we
need to ensure that it still lays within the document.